requirements.txt
amqp==5.2.0
appdirs==1.4.4
asgiref==3.8.1
atomicwrites==1.4.0
backcall==0.2.0
backports.weakref==1.0.post1
billiard==4.2.1
celery==5.4.0
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
clyent==1.2.1
distlib==0.3.8
dj-database-url==2.2.0
Django==5.0.4
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docopt==0.6.2
et-xmlfile==1.1.0
fonttools==4.25.0
gunicorn==23.0.0
html5==0.0.9
inflection==0.5.1
ipython==8.12.3
jsonpointer==2.1
kombu==5.4.2
munkres==1.1.4
nbconvert==7.16.4
nodejs==0.1.1
openpyxl==3.0.10
optional-django==0.1.0
panda==0.3.1
patsy==0.5.3
pipreqs==0.5.0
ply==3.11
protobuf==3.20.3
psycopg2==2.9.9
pyasn1-modules==0.2.8
pycurl==7.45.2
PyDispatcher==2.0.5
pygame==2.5.2
pyls-spyder==0.4.0
PyQt5==5.15.10
PyQtWebEngine==5.15.6
python-lsp-jsonrpc==1.0.0
python-dotenv
redis==5.1.1
scipy==1.11.4
sqlparse==0.5.0
vine==5.1.0
virtualenv==20.26.1
webencodings==0.5.1
yarg==0.1.9
zstandard==0.19.0
